Only add real generated outputs (#365)
Fixes #362

We are currently adding outputs that won't get created because they
match build actions in the wrong package.

Change the assert to an `if/throw` since we never want to allow this
even in non-checked mode.

+++ build_runner/lib/src/asset/file_based.dart
@@ -96,7 +96,10 @@ class FileBasedAssetWriter implements RunnerAssetWriter {
 
   @override
   Future delete(AssetId id) {
-    assert(id.package == packageGraph.root.name);
+    if (id.package != packageGraph.root.name) {
+      throw new InvalidOutputException(
+          id, 'Should not delete assets outside of ${packageGraph.root.name}');
+    }
     if (onDelete != null) onDelete(id);
 
     var file = _fileFor(id, packageGraph);
